pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rauq"
version = "0.1.0"
description = "RAUQ uncertainty benchmarking suite"
authors = [{name = "Codex"}]
requires-python = ">=3.11"
dependencies = [
  "torch>=2.2",
  "transformers>=4.44",
  "datasets>=2.16",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
tests = ["pytest>=7.4"]
plots = ["alignscore==0.1.1", "comet-ml>=3.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
